+++ stories/components/form-field/form-field.stories.ts
@@ -200,8 +200,8 @@ const PrefixSuffixGroupTemplate: StoryFn<Partial<FormFieldStoriesParams>> = (arg
             <wcs-select-option>Japan</wcs-select-option>
         </wcs-select>
         <wcs-input placeholder="Region"></wcs-input>
-        <wcs-button shape="square" slot="suffix" ripple="false">
-            <wcs-mat-icon icon="search" role="img" aria-label="Search"></wcs-mat-icon>
+        <wcs-button shape="square" slot="suffix">
+            <wcs-mat-icon icon="clear" role="img" aria-label="Clear input"></wcs-mat-icon>
         </wcs-button>
     </wcs-form-field>